add failing test for @javax.enterprise.inject.Produces @Resource(lookup = "java:jboss/cassandradriver/test")

+++ testsuite/nosql/src/test/java/org/jboss/as/test/compat/nosql/cassandra/jaxrs/ClientResource.java
@@ -27,6 +27,7 @@ import java.util.concurrent.ExecutionException;
 import javax.annotation.Resource;
 import javax.ejb.Stateless;
 import javax.ws.rs.GET;
+import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 
@@ -45,7 +46,12 @@ public class ClientResource {
     // can only use @Resource in EE components, which is why this is a stateless session bean.
     @Resource(lookup = "java:jboss/cassandradriver/test")
     private Cluster cluster;
+
     private Session session;
+
+    @javax.enterprise.inject.Produces
+    @Resource(lookup = "java:jboss/cassandradriver/test")
+    private static Cluster injectedCluster;
 
     @GET
     @Produces({"text/plain"})
@@ -69,7 +75,31 @@ public class ClientResource {
 
     }
 
+    @POST
+    @Produces({"text/plain"})
+    public String getInjectedConnection() {
+        openInjectedConnection();
+        try {
+            session.execute("CREATE TABLE journal (name varchar primary key, when date, comment varchar)");
+            session.execute("INSERT INTO journal (name, when, comment) VALUES " +
+                    "('Scott Marlow','2016-09-05','try injected connection')");
+            ResultSetFuture results = session.executeAsync("SELECT JSON * FROM journal");
+            return results.get().one().getString("[json]");
+        } catch (InterruptedException e) {
+            throw new RuntimeException("Interrupted while getting results",e);
+        } catch (ExecutionException e) {
+            throw new RuntimeException("Failure while getting results", e);
+        } finally {
+            closeConnection();
+        }
+
+    }
+
+
     private void openConnection() {
+        if(cluster == null) {
+            throw new RuntimeException("failed to get connection to NoSQL database using @javax.annotation.Resource");
+        }
         session = cluster.connect();
         session.execute("CREATE KEYSPACE testspace WITH REPLICATION = { 'class' : 'SimpleStrategy', 'replication_factor' : 1 };");
         session = cluster.connect("testspace");
@@ -77,9 +107,20 @@ public class ClientResource {
 
 
     private void closeConnection() {
-        session.execute("DROP KEYSPACE testspace");
-        session.close();
-        session = null;
+        if (session != null) {
+            session.execute("DROP KEYSPACE testspace");
+            session.close();
+            session = null;
+        }
+    }
+
+    private void openInjectedConnection() {
+        if(injectedCluster == null) {
+            throw new RuntimeException("failed to get connection to NoSQL database using @javax.enterprise.inject.Produces @Resource");
+        }
+        session = injectedCluster.connect();
+        session.execute("CREATE KEYSPACE testspace WITH REPLICATION = { 'class' : 'SimpleStrategy', 'replication_factor' : 1 };");
+        session = cluster.connect("testspace");
     }
 
 }
